sstable: split Iterator and helpers out of reader.go
